feat(api): extract webhook/health handlers and add service readiness check
Move handle_provider_webhook out of the monolithic lib.rs into
handlers/webhook.rs and the four health-check functions into
handlers/health.rs, resolving the structural inconsistency noted in
spec review (issue #12).

Add ServiceHealthChecker that verifies at least one webhook provider is
registered before reporting ready (/ready), replacing the always-true
DefaultHealthChecker stub used in production (issue #11). liveness
(/health/live) continues to return 'alive' unconditionally.

Also add ProviderRegistry::is_empty/len helper methods used by the new
checker, and update start_server to accept Arc<ProviderRegistry> so the
health checker and AppState share the same registry instance.

(auto via agent)
