+++ crates/queue-keeper-service/src/main.rs
@@ -15,7 +15,7 @@ mod signature_validator;
 
 use circuit_breaker::queue::{CircuitBreakerQueueClient, CircuitBreakerQueueProvider};
 use queue_keeper_api::{
-    start_server, DefaultEventStore, DefaultHealthChecker, ProviderId, ProviderRegistry,
+    start_server, DefaultEventStore, ServiceHealthChecker, ProviderId, ProviderRegistry,
     QueueBackendConfig, ServiceConfig, ServiceError,
 };
 use queue_keeper_core::adapters::{memory_key_vault::InMemorySecretCache, AzureKeyVaultProvider};
@@ -253,7 +253,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
     }
 
-    let health_checker = Arc::new(DefaultHealthChecker);
+    let health_checker = Arc::new(ServiceHealthChecker::new(Arc::new(provider_registry.clone())));
     let event_store = Arc::new(DefaultEventStore);
 
     // -------------------------------------------------------------------------
@@ -295,7 +295,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Start the server
     if let Err(e) = start_server(
         service_config,
-        provider_registry,
+        Arc::new(provider_registry),
         health_checker,
         event_store,
         generic_provider_ids,
